SMTChecker: Fix error when initializing fixed-sized-bytes array
When encoding initialization of an array of fixed-sized bytes, we need
to make sure the elements of the inlined array have the right SMT sort.
Because of implict conversion from string literals to fixed-sized bytes,
we need to pass the information about the desired type when encoding the
individual elements.