SMTChecker: Fix crash in BMC engine regarding state variables
Previously, analyzing a call to a getter to a contract then has not been
analyzed yet with BMC would result in a crash because BMC would not know
about the state variable being accessed.

To fix this, we let BMC know about all state variables in all contracts
during initialization.