SMTChecker: Add support for block.blobbasefee and blobhash()
We overapproximate the behaviour here in the same way as for
`block.basefee` and `blockhash`.
For the first we only add the basic domain constraints and the second is
modelled as an uninterpreted function.

We also model the fact that the number of blobs in transaction is
limited by returning 0 from blobhash() if the argument is greater or
equal to the limit.
The current limit is 6, but it will soon be increased to 9 in pectra
update.
Since we overapproximate the behaviour anyway, we can use the pectra
limit immediately.